+++ tests_e2e/conftest.py
@@ -3,6 +3,7 @@ import subprocess
 import tempfile
 from pathlib import Path
 from tempfile import TemporaryDirectory
+from typing import List
 
 import pytest
 from snowcli import __about__
@@ -60,14 +61,14 @@ def _build_snowcli(venv_path: Path, test_root_path: Path) -> None:
 def _install_snowcli_with_external_plugin(
     venv_path: Path, test_root_path: Path
 ) -> None:
-    version = __about__.VERSION.replace("-", ".")
+    whl_file = _find_whl_file(os.listdir(test_root_path / f"../dist/"))
     subprocess.check_call(
         [
             _python_path(venv_path),
             "-m",
             "pip",
             "install",
-            test_root_path / f"../dist/snowflake_cli_labs-{version}0-py3-none-any.whl",
+            test_root_path / f"../dist/{whl_file}",
         ]
     )
     subprocess.check_call(
@@ -86,3 +87,10 @@ def _install_snowcli_with_external_plugin(
 
 def _python_path(venv_path: Path) -> Path:
     return venv_path / "bin" / "python"
+
+
+def _find_whl_file(files: List[str]) -> str:
+    for file in files:
+        if file.endswith(".whl"):
+            return file
+    raise RuntimeError(f"Not found whl file in: {files}")